remove dead code after INST_ARCHLIBDIR install dest change
Previously for cperl, Config went into the "default" location but for
CPAN went to ARCHLIB, in commit "Config: install into ARCHLIB not LIB"
Reini changed it so both cperl and CPAN goto ARCHLIB, the conditional
can now be removed. Move the comment also so no future maintainers think
it is a good idea to put Config.pm in non-ARCH lib.